Nest: Expand rte_src.private_id to u64
In general, private_id is sparse and protocols may want to map some
internal values directly into it. For example, L3VPN needs to
map VPN route discriminators to private_id.

OTOH, u32 is enough for global_id, as these identifiers are dense.

+++ nest/route.h
@@ -439,7 +439,7 @@ struct nexthop {
 struct rte_src {
   struct rte_src *next;			/* Hash chain */
   struct proto *proto;			/* Protocol the source is based on */
-  u32 private_id;			/* Private ID, assigned by the protocol */
+  u64 private_id;			/* Private ID, assigned by the protocol */
   u32 global_id;			/* Globally unique ID of the source */
   unsigned uc;				/* Use count */
 };
